+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 4.0.3 (UNRELEASED)
 * Fixed an issue that prevented the UI from loading on servers with a `RewriteOptions inherit` configuration.
+* Fixed `$format_lang_names` setting to work with the new UI.
 
 ## 4.0.2 (November 11, 2025)
 * Added Italian translations.