+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fixed `$format_lang_names` setting to work with the new UI.
 * Fixed a geocoding issue -- if the county and/or zip code is automatically computed by the geocoder, we don't want to feed old and possibly incorrect values to the geocoder if the user edits an address.
 * Fixed problems with format key validation -- server now checks for duplicate keys correctly, and disallows changes to the English keys for the reserved formats HY, TC, and VM.
+* Added Portuguese translations.
 
 ## 4.0.2 (November 11, 2025)
 * Added Italian translations.